requirements.txt
torch
torchvision
Pillow
flask
flask-cors
